初步完成mybatis-plus基础配置内容

+++ common-cache-starter/src/main/java/com/lwh147/common/cache/properties/JetCacheProperties.java
@@ -2,6 +2,7 @@ package com.lwh147.common.cache.properties;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.data.redis.core.RedisTemplate;
 
 /**
  * JetCache配置类
@@ -13,31 +14,31 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
 @ConfigurationProperties(prefix = "jetcache")
 public class JetCacheProperties {
     /**
-     * 数据库，默认1号，与RedisTemplate区分
+     * 数据库，默认为 {@code 1}，与 {@link RedisTemplate} 区分
      **/
     private Integer database = 1;
     /**
-     * 本地缓存数量限制，默认1024个
+     * 本地缓存数量限制，默认 {@code 1024} 个
      **/
     private Integer localLimit = 1024;
     /**
-     * 本地缓存过期时间，默认60s
+     * 本地缓存过期时间，单位秒，默认 {@code 60}
      **/
-    private Integer localExpiredIn = 60;
+    private Long localExpiredIn = 60L;
     /**
-     * 远程缓存过期时间，默认180s
+     * 远程缓存过期时间，单位秒，默认 {@code 180}
      **/
-    private Integer remoteExpiredIn = 180;
+    private Long remoteExpiredIn = 180L;
     /**
-     * 缓存使用情况统计时间间隔，默认 15 分钟
+     * 缓存使用情况统计时间间隔，单位分钟，默认 {@code 15}
      **/
     private Integer statIntervalMinutes = 15;
     /**
-     * 缓存刷新时间间隔，默认60s
+     * 缓存刷新时间间隔，单位秒，默认 {@code 60}
      **/
     private Integer refresh = 60;
     /**
-     * 停止刷新缓存的等待时间，默认180s
+     * 停止刷新缓存的等待时间，单位秒，默认 {@code 180}
      **/
     private Integer stopRefreshAfterLastAccess = 180;
 }
